# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e2456320cfc6233ae9336d9
  title: Manage_Film_Example
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  Actor  -----
  /actor/:
    post:
      x-skaffolder-id: 5e2456330cfc6233ae93370a
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336f4
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Actor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                birthDate:
                  type: integer
                name:
                  type: string
                surname:
                  type: string
    get:
      x-skaffolder-id: 5e2456330cfc6233ae933711
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336f4
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Actor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Actor"
  /actor/{id}:
    delete:
      x-skaffolder-id: 5e2456330cfc6233ae933710
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336f4
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Actor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e2456330cfc6233ae93370b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336f4
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Actor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
    post:
      x-skaffolder-id: 5e2456330cfc6233ae933708
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336f4
      x-skaffolder-resource: Actor
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Actor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Actor"
    
  #-----  Film  -----
...
  /user/:
    post:
      x-skaffolder-id: 5e2456320cfc6233ae9336e9
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336e4
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5e2456320cfc6233ae9336eb
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e2456320cfc6233ae9336e4
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
    Actor:
      x-skaffolder-id: 5e2456320cfc6233ae9336f4
      x-skaffolder-id-db: 5e2456320cfc6233ae9336da
      x-skaffolder-id-entity: 5e2456320cfc6233ae9336f0
      x-skaffolder-url: /actor
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        birthDate:
          type: integer
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336f3
          x-skaffolder-type: Date
        name:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336f1
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336f2
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
        - name
...
    User:
      x-skaffolder-id: 5e2456320cfc6233ae9336e4
      x-skaffolder-id-db: 5e2456320cfc6233ae9336da
      x-skaffolder-id-entity: 5e2456320cfc6233ae9336dc
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336df
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336e0
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336de
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336e2
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336e1
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5e2456320cfc6233ae9336dd
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...